Improve project structure folders
