+++ src/components/authors/FormAuthor.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { useForm } from "react-hook-form";
-import { addBook } from "../services/indexedDB";
+import { addAuthor } from "../../services/indexedDB";
 import { useState } from "react";
 
 const Form = styled.form`
@@ -9,24 +9,6 @@ const Form = styled.form`
 `;
 
 const Input = styled.input`
-  font-family: "Open Sans", sans-serif;
-  font-size: 16px;
-  font-weight: 500;
-  font-style: normal;
-  color: #333;
-  background-color: #fff;
-  height: 40px;
-  border: 1px solid #aaa;
-  border-radius: 3px;
-  padding: 0px 10px;
-  margin-top: 15px;
-  transition: all 0.3s;
-  &:focus {
-    background-color: #ddd;
-  }
-`;
-
-const Select = styled.select`
   font-family: "Open Sans", sans-serif;
   font-size: 16px;
   font-weight: 500;
@@ -87,7 +69,7 @@ const SubmitButton = styled.button`
   }
 `;
 
-const FormBook = ({ onAddData, authorsData }) => {
+const FormAuthor = ({ onAddData }) => {
   const {
     register,
     handleSubmit,
@@ -103,10 +85,10 @@ const FormBook = ({ onAddData, authorsData }) => {
     setSubmitError("");
 
     try {
-      const newBook = await addBook(data);
+      const newAuthor = await addAuthor(data);
 
-      setSubmitSuccess("Livro cadastrado com sucesso!");
-      onAddData(newBook);
+      setSubmitSuccess("Autor cadastrado com sucesso!");
+      onAddData(newAuthor);
       reset();
     } catch (error) {
       setSubmitError(error);
@@ -127,23 +109,7 @@ const FormBook = ({ onAddData, authorsData }) => {
       />
       {errors.name && <InputError>{errors.name.message}</InputError>}
 
-      <Select
-        {...register("author", { required: "Campo obrigatório" })}
-        defaultValue=""
-      >
-        <option value="" disabled>
-          Autor
-        </option>
-        {authorsData &&
-          authorsData.map((author) => (
-            <option key={author.id} value={author.id}>
-              {author.name}
-            </option>
-          ))}
-      </Select>
-      {errors.author && <InputError>{errors.author.message}</InputError>}
-
-      <Input type="number" {...register("pages")} placeholder="Páginas" />
+      <Input {...register("email")} placeholder="E-mail" type="email" />
 
       <SubmitButton
         onClick={handleClick}
@@ -159,4 +125,4 @@ const FormBook = ({ onAddData, authorsData }) => {
   );
 };
 
-export default FormBook;
+export default FormAuthor;
